+++ NEWS.md
@@ -4,6 +4,8 @@
 
 ## New features
 
+
+* Added `process_jags_samples()` and `plot_predicted_curve()`  (#68)
 * Changes to `run_mod()` (#79):
    - `jags.post` now optionally included in output, as specified by argument
    `with_post`